requirements.txt
meshtastic>=1.2.0
requests>=2.28.0
PyPubSub>=4.0.0
bleak>=0.20.0
packaging>=21.0
protobuf>=3.20.0
pyserial>=3.5
pyyaml>=6.0
tabulate>=0.8.0 